# Improve performance of action's processing