--- a/src/testParser.ts
+++ b/src/testParser.ts
@@ -110,7 +110,12 @@ function safeParseInt(line: string | null): number | null {
  * Modification Copyright 2022 Mike Penz
  * https://github.com/mikepenz/action-junit-report/
  */
+const resolvePathCache: {[key: string]: string} = {}
 export async function resolvePath(fileName: string, excludeSources: string[], followSymlink = false): Promise<string> {
+  if (resolvePathCache[fileName]) {
+    return resolvePathCache[fileName]
+  }
+
   core.debug(`Resolving path for ${fileName}`)
   const normalizedFilename = fileName.replace(/^\.\//, '') // strip relative prefix (./)
   const globber = await glob.create(`**/${normalizedFilename}.*`, {
@@ -124,10 +129,12 @@ export async function resolvePath(fileName: string, excludeSources: string[], fo
     if (!found) {
       const path = result.slice(searchPath.length + 1)
       core.debug(`Resolved path: ${path}`)
-      return path
+      resolvePathCache[fileName] = path
+      return resolvePathCache[fileName]
     }
   }
-  return normalizedFilename
+  resolvePathCache[fileName] = normalizedFilename
+  return resolvePathCache[fileName]
 }
 
 /**