--- a/src/testParser.ts
+++ b/src/testParser.ts
@@ -307,6 +307,17 @@ async function parseSuite(
         testcase.skipped || testcase._attributes.status === 'disabled' || testcase._attributes.status === 'ignored'
       const failed = testFailure && !skip // test faiure, but was skipped -> don't fail if a ignored test failed
       const success = !testFailure // not a failure -> thus a success
+      const annotationLevel = success || skip ? 'notice' : 'failure' // a skipped test shall not fail the run
+
+      if (skip) {
+        skipped++
+      }
+
+      // If this won't be reported as a failure and processing all passed tests
+      // isn't enabled, then skip the rest of the processing.
+      if (annotationLevel !== 'failure' && !annotatePassed) {
+        continue
+      }
 
       // in some definitions `failure` may be an array
       const failures = testcase.failure
@@ -317,9 +328,6 @@ async function parseSuite(
       // the action only supports 1 failure per testcase
       const failure = failures ? failures[0] : undefined
 
-      if (skip) {
-        skipped++
-      }
       const stackTrace: string = (
         (failure && failure._cdata) ||
         (failure && failure._text) ||
@@ -403,7 +411,7 @@ async function parseSuite(
         end_line: pos.line,
         start_column: 0,
         end_column: 0,
-        annotation_level: success || skip ? 'notice' : 'failure', // a skipped test shall not fail the run
+        annotation_level: annotationLevel,
         status: skip ? 'skipped' : success ? 'success' : 'failure',
         title: escapeEmoji(title),
         message: escapeEmoji(message),