… dist/index.js
… dist/index.js.map