@mikepenz 2024-06-22
Unfortunately I had to reverse the second commit for v4 of the action. It will be included in v5 as to make it work consistent it will result in some small behavior changes. 

thanks again for the great contribution!